# Conflicts:
#	e2e/gui/src/test/java/com/epam/pipeline/autotests/ao/Primitive.java
#	e2e/gui/src/test/java/com/epam/pipeline/autotests/ao/StorageContentAO.java
#	e2e/gui/src/test/java/com/epam/pipeline/autotests/ao/SystemManagementAO.java

+++ scripts/autoscaling/init_multicloud_v1.15.4.sh
@@ -1,14 +1,22 @@
 #!/bin/bash
 
+_API_URL="@API_URL@"
+_API_TOKEN="@API_TOKEN@"
+
 launch_token="/etc/user_data_launched"
 if [[ -f "$launch_token" ]]; then exit 0; fi
 
 user_data_log="/var/log/user_data.log"
 exec > "$user_data_log" 2>&1
 
-function check_installed {
-    command -v "$1" >/dev/null 2>&1
-    return $?
+function check_gpu_available {
+    command -v "nvidia-smi" >/dev/null 2>&1
+    if [ $? -ne 0 ]; then echo "nvidia-smi not found"; return 1; fi
+    
+    nvidia-smi &> /dev/null
+    if [ $? -eq 9 ]; then echo "NVIDIA driver is not loaded"; return 1; fi
+    
+    return 0
 }
 
 function update_nameserver {
@@ -40,6 +48,46 @@ function update_nameserver {
     echo "nameserver $nameserver" >> /etc/resolv.conf
     chattr +i /etc/resolv.conf
   fi
+}
+
+function wait_device_uuid {
+    local device_name="$1"
+    local attempts="$2"
+
+    export DEVICE_UUID=""
+    echo "Waiting for device uuid $device_name..."
+    for i in $(seq 1 "$attempts"); do
+        export DEVICE_UUID="$(lsblk -sdrpn -o NAME,UUID | awk -F'[ ]' '$1 ~ device_name { print $2 }' device_name="$device_name" | head -n 1)"
+        if [ "$DEVICE_UUID" ]; then
+            echo "Device uuid is ready $device_name ($DEVICE_UUID)"
+            return 0
+        fi
+        echo "Waiting for device uuid..."
+        sleep 1
+    done
+
+    echo "Device uuid is NOT ready after $attempts seconds"
+    return 1
+}
+
+function wait_device_part {
+    local device_prefix="$1"
+    local attempts="$2"
+
+    export DEVICE_NAME=""
+    echo "Waiting for device part $device_prefix..."
+    for i in $(seq 1 "$attempts"); do
+        export DEVICE_NAME="$(lsblk -sdrpn -o NAME,TYPE,MOUNTPOINT | awk -F'[ ]' '$1 ~ device_prefix && $2 == "part" && $3 == "" { print $1 }' device_prefix="$device_prefix" | head -n 1)"
+        if [ "$DEVICE_NAME" ]; then
+            echo "Device part is ready $DEVICE_NAME"
+            return 0
+        fi
+        echo "Waiting for device part..."
+        sleep 1
+    done
+
+    echo "Device part is NOT ready after $attempts seconds"
+    return 1
 }
 
 function setup_swap_device {
@@ -77,9 +125,47 @@ function setup_swap_device {
             echo "Unable to swapon at $swap_drive_name"
             return 1
         fi
-        swap_drive_uuid=$(lsblk -sdrpn -o NAME,UUID | awk '$1 == "'"$swap_drive_name"'" { print $2 }')
-        echo "UUID=$swap_drive_uuid none swap sw 0 0" >> /etc/fstab
+        wait_device_uuid "$swap_drive_name" 10
+        echo "UUID=$DEVICE_UUID none swap sw 0 0" >> /etc/fstab
     fi
+}
+
+function enable_emergency_termination_service() {
+  _EMERGENCY_TERMINATOR_PRESENT=0
+  _CURRENT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
+  if [ -f "$_CURRENT_DIR/emergency_node_terminator" ]; then
+    cp "$_CURRENT_DIR/emergency_node_terminator" "/usr/bin/emergency_node_terminator"
+    _EMERGENCY_TERMINATOR_PRESENT=1
+  else
+    _EMERGENCY_TERMINATOR__URL="$(dirname $_API_URL)/emergency_node_terminator.sh"
+    echo "Cannot find $_CURRENT_DIR/emergency_node_terminator, downloading from ${_EMERGENCY_TERMINATOR__URL}"
+    curl -skf "${_EMERGENCY_TERMINATOR__URL}" > /usr/bin/emergency_node_terminator
+    if [ $? -ne 0 ]; then
+      echo "Error while downloading emergency_node_terminator script"
+    else
+      _EMERGENCY_TERMINATOR_PRESENT=1
+    fi
+  fi
+
+  if [ $_EMERGENCY_TERMINATOR_PRESENT -eq 1 ]; then
+    chmod +x /usr/bin/emergency_node_terminator
+  cat >/etc/systemd/system/emergency_node_terminator.service <<EOL
+[Unit]
+Description=Cloud Pipeline Emergency Node Termination Daemon
+Documentation=https://cloud-pipeline.com/
+
+[Service]
+Restart=always
+StartLimitInterval=0
+RestartSec=10
+ExecStart=/usr/bin/emergency_node_terminator "$_API_URL" "$_API_TOKEN" "$_KUBE_NODE_NAME"
+
+[Install]
+WantedBy=multi-user.target
+EOL
+    systemctl enable emergency_node_terminator
+    systemctl start emergency_node_terminator
+  fi
 }
 
 GLOBAL_DISTRIBUTION_URL="@GLOBAL_DISTRIBUTION_URL@"
@@ -88,12 +174,27 @@ if [ ! "$GLOBAL_DISTRIBUTION_URL" ] || [[ "$GLOBAL_DISTRIBUTION_URL" == "@"*"@" 
 fi
 export GLOBAL_DISTRIBUTION_URL
 
+_WO="--timeout=10 --waitretry=1 --tries=10"
+wget $_WO "${GLOBAL_DISTRIBUTION_URL}tools/nvme-cli/1.16/nvme.gz" -O /bin/nvme.gz && \
+gzip -d /bin/nvme.gz && \
+chmod +x /bin/nvme
+
+@custom_script_pre@
+
 swap_size="@swap_size@"
 setup_swap_device "${swap_size:-0}"
 
 FS_TYPE="@FS_TYPE@"
 
+_ds=()
+if [[ -x /bin/nvme ]]; then
+  _ds=($(nvme list | grep 'Instance Storage' | awk '{ print $1 }'))
+fi
 UNMOUNTED_DRIVES=($(lsblk -sdrpn -o NAME,TYPE,MOUNTPOINT | awk '$2 == "disk" && $3 == "" { print $1 }'))
+_dsc=( $({ printf "%s\n" "${UNMOUNTED_DRIVES[@]}" | sort -u; printf "%s\n" "${_ds[@]}" "${_ds[@]}"; } | sort | uniq -u) )
+if [[ ${#_dsc[@]} > 0 ]]; then
+  UNMOUNTED_DRIVES=("${_dsc[@]}")
+fi
 if [[ ${#UNMOUNTED_DRIVES[@]} > 1 ]]; then
   pvcreate ${UNMOUNTED_DRIVES[@]}
   vgcreate nvmevg ${UNMOUNTED_DRIVES[@]}
@@ -104,13 +205,14 @@ elif [[ ${#UNMOUNTED_DRIVES[@]} == 1 ]]; then
   PARTITION_RESULT=$(sfdisk -d $DRIVE_NAME 2>&1)
   if [[ $PARTITION_RESULT == "" ]]; then
       (echo o; echo n; echo p; echo; echo; echo; echo w) | fdisk $DRIVE_NAME
-      DRIVE_NAME="${DRIVE_NAME}1"
+      wait_device_part "$DRIVE_NAME" 10
+      DRIVE_NAME="$DEVICE_NAME"
   elif [[ $PARTITION_RESULT == *"No such device or address"* ]]; then
       echo "Cannot create partition for ${DRIVE_NAME}, falling back to root volume"
       unset DRIVE_NAME
   fi
 else
-  echo "No unmounted drives found. Root volume is used as for the /ebs"
+  echo "No unmounted drives found. Root volume is used for the /ebs"
 fi
 
 MOUNT_POINT="/ebs"
@@ -140,42 +242,58 @@ fi
 
 systemctl stop docker
 
-_DOCKER_SYS_IMGS="/ebs/docker-system-images"
-rm -rf $_DOCKER_SYS_IMGS
-_KUBE_SYSTEM_PODS_DISTR="@SYSTEM_PODS_DISTR_PREFIX@"
-if [ ! "$_KUBE_SYSTEM_PODS_DISTR" ] || [[ "$_KUBE_SYSTEM_PODS_DISTR" == "@"*"@" ]]; then
-  _KUBE_SYSTEM_PODS_DISTR="${GLOBAL_DISTRIBUTION_URL}tools/kube/1.15.4/docker"
+skip_system_images_load="@SKIP_SYSTEM_IMAGES_LOAD@"
+if [ "$skip_system_images_load" != "true" ]; then
+  _DOCKER_SYS_IMGS="/ebs/docker-system-images"
+  rm -rf $_DOCKER_SYS_IMGS
+  _KUBE_SYSTEM_PODS_DISTR="@SYSTEM_PODS_DISTR_PREFIX@"
+  if [ ! "$_KUBE_SYSTEM_PODS_DISTR" ] || [[ "$_KUBE_SYSTEM_PODS_DISTR" == "@"*"@" ]]; then
+    _KUBE_SYSTEM_PODS_DISTR="${GLOBAL_DISTRIBUTION_URL}tools/kube/1.15.4/docker"
+  fi
+  mkdir -p $_DOCKER_SYS_IMGS
+  wget $_WO "${_KUBE_SYSTEM_PODS_DISTR}/calico-node-v3.14.1.tar" -O $_DOCKER_SYS_IMGS/calico-node-v3.14.1.tar && \
+  wget $_WO "${_KUBE_SYSTEM_PODS_DISTR}/calico-pod2daemon-flexvol-v3.14.1.tar" -O $_DOCKER_SYS_IMGS/calico-pod2daemon-flexvol-v3.14.1.tar &&
+  wget $_WO "${_KUBE_SYSTEM_PODS_DISTR}/calico-cni-v3.14.1.tar" -O $_DOCKER_SYS_IMGS/calico-cni-v3.14.1.tar && \
+  wget $_WO "${_KUBE_SYSTEM_PODS_DISTR}/k8s.gcr.io-kube-proxy-v1.15.4.tar" -O $_DOCKER_SYS_IMGS/k8s.gcr.io-kube-proxy-v1.15.4.tar && \
+  wget $_WO "${_KUBE_SYSTEM_PODS_DISTR}/quay.io-coreos-flannel-v0.11.0.tar" -O $_DOCKER_SYS_IMGS/quay.io-coreos-flannel-v0.11.0.tar && \
+  wget $_WO "${_KUBE_SYSTEM_PODS_DISTR}/k8s.gcr.io-pause-3.1.tar" -O $_DOCKER_SYS_IMGS/k8s.gcr.io-pause-3.1.tar
+  if [ $? -ne 0 ]; then
+    _DOCKER_SYS_IMGS="/opt/docker-system-images"
+  fi
 fi
-mkdir -p $_DOCKER_SYS_IMGS
-_WO="--timeout=10 --waitretry=1 --tries=10"
-wget $_WO "${_KUBE_SYSTEM_PODS_DISTR}/calico-node-v3.14.1.tar" -O $_DOCKER_SYS_IMGS/calico-node-v3.14.1.tar && \
-wget $_WO "${_KUBE_SYSTEM_PODS_DISTR}/calico-pod2daemon-flexvol-v3.14.1.tar" -O $_DOCKER_SYS_IMGS/calico-pod2daemon-flexvol-v3.14.1.tar &&
-wget $_WO "${_KUBE_SYSTEM_PODS_DISTR}/calico-cni-v3.14.1.tar" -O $_DOCKER_SYS_IMGS/calico-cni-v3.14.1.tar && \
-wget $_WO "${_KUBE_SYSTEM_PODS_DISTR}/k8s.gcr.io-kube-proxy-v1.15.4.tar" -O $_DOCKER_SYS_IMGS/k8s.gcr.io-kube-proxy-v1.15.4.tar && \
-wget $_WO "${_KUBE_SYSTEM_PODS_DISTR}/quay.io-coreos-flannel-v0.11.0.tar" -O $_DOCKER_SYS_IMGS/quay.io-coreos-flannel-v0.11.0.tar && \
-wget $_WO "${_KUBE_SYSTEM_PODS_DISTR}/k8s.gcr.io-pause-3.1.tar" -O $_DOCKER_SYS_IMGS/k8s.gcr.io-pause-3.1.tar
-if [ $? -ne 0 ]; then
-  _DOCKER_SYS_IMGS="/opt/docker-system-images"
+
+
+if [ ! -x /bin/wondershaper ]; then
+  wget $_WO "${GLOBAL_DISTRIBUTION_URL}tools/wondershaper/wondershaper" -O /bin/wondershaper
+  chmod +x /bin/wondershaper
 fi
 
 mkdir -p /etc/docker
 
-if [[ $FS_TYPE == "ext4" ]]; then
-  DOCKER_STORAGE_DRIVER="overlay2"
-  DOCKER_STORAGE_OPTS='"storage-opts": ["overlay2.override_kernel_check=true"],'
-else
-  DOCKER_STORAGE_DRIVER="btrfs"
-  DOCKER_STORAGE_OPTS=""
+docker_storage_driver="@DOCKER_STORAGE_DRIVER@"
+if [ -z "$docker_storage_driver" ] || [[ "$docker_storage_driver" == "@"*"@" ]]; then
+  if [[ $FS_TYPE == "ext4" ]]; then
+    docker_storage_driver="overlay2"
+    DOCKER_STORAGE_OPTS='"storage-opts": ["overlay2.override_kernel_check=true"],'
+  else
+    docker_storage_driver="btrfs"
+    DOCKER_STORAGE_OPTS=""
+  fi
 fi
 
-if check_installed "nvidia-smi"; then
+docker_data_root="@DOCKER_DATA_ROOT@"
+if [ -z "$docker_data_root" ] || [[ "$docker_data_root" == "@"*"@" ]]; then
+  docker_data_root="/ebs/docker"
+fi
+
+if check_gpu_available; then
   nvidia-persistenced --persistence-mode
 
 cat <<EOT > /etc/docker/daemon.json
 {
   "exec-opts": ["native.cgroupdriver=systemd"],
-  "data-root": "/ebs/docker",
-  "storage-driver": "$DOCKER_STORAGE_DRIVER",
+  "data-root": "$docker_data_root",
+  "storage-driver": "$docker_storage_driver",
   $DOCKER_STORAGE_OPTS
   "max-concurrent-uploads": 1,
   "default-runtime": "nvidia",
@@ -191,8 +309,8 @@ else
 cat <<EOT > /etc/docker/daemon.json
 {
   "exec-opts": ["native.cgroupdriver=systemd"],
-  "data-root": "/ebs/docker",
-  "storage-driver": "$DOCKER_STORAGE_DRIVER",
+  "data-root": "$docker_data_root",
+  "storage-driver": "$docker_storage_driver",
   $DOCKER_STORAGE_OPTS
   "max-concurrent-uploads": 1
 }
@@ -219,20 +337,22 @@ modprobe nfsd
 
 chmod +x /etc/rc.d/rc.local
 
-cloud=$(curl --head -s http://169.254.169.254/latest/dynamic/instance-identity/document | grep Server | cut -f2 -d:)
+_CLOUD_TOKEN=$(curl -s -X PUT "http://169.254.169.254/latest/api/token" -H "X-aws-ec2-metadata-token-ttl-seconds: 21600")
+cloud=$(curl --head -s -H "X-aws-ec2-metadata-token: $_CLOUD_TOKEN" http://169.254.169.254/latest/dynamic/instance-identity/document | grep Server | cut -f2 -d:)
 gcloud_header=$(curl --head -s http://169.254.169.254/latest/dynamic/instance-identity/document | grep Metadata-Flavor | cut -f2 -d:)
 
 if [[ $cloud == *"EC2"* ]]; then
-    _CLOUD_REGION=$(curl -s http://169.254.169.254/latest/dynamic/instance-identity/document | grep region | cut -d\" -f4)
-    _CLOUD_INSTANCE_AZ=$(curl -s http://169.254.169.254/latest/dynamic/instance-identity/document | grep availabilityZone | cut -d\" -f4)
-    _CLOUD_INSTANCE_ID=$(curl -s http://169.254.169.254/latest/dynamic/instance-identity/document | grep instanceId | cut -d\" -f4)
-    _CLOUD_INSTANCE_TYPE=$(curl -s http://169.254.169.254/latest/dynamic/instance-identity/document | grep instanceType | cut -d\" -f4)
-    _CLOUD_INSTANCE_IMAGE_ID=$(curl -s http://169.254.169.254/latest/dynamic/instance-identity/document | grep imageId | cut -d\" -f4)
-    _CI_IP=$(curl -s http://169.254.169.254/latest/meta-data/local-ipv4)
+    _CLOUD_REGION=$(curl -s -H "X-aws-ec2-metadata-token: $_CLOUD_TOKEN" http://169.254.169.254/latest/dynamic/instance-identity/document | grep region | cut -d\" -f4)
+    _CLOUD_INSTANCE_AZ=$(curl -s -H "X-aws-ec2-metadata-token: $_CLOUD_TOKEN" http://169.254.169.254/latest/dynamic/instance-identity/document | grep availabilityZone | cut -d\" -f4)
+    _CLOUD_INSTANCE_ID=$(curl -s -H "X-aws-ec2-metadata-token: $_CLOUD_TOKEN" http://169.254.169.254/latest/dynamic/instance-identity/document | grep instanceId | cut -d\" -f4)
+    _CLOUD_INSTANCE_TYPE=$(curl -s -H "X-aws-ec2-metadata-token: $_CLOUD_TOKEN" http://169.254.169.254/latest/dynamic/instance-identity/document | grep instanceType | cut -d\" -f4)
+    _CLOUD_INSTANCE_IMAGE_ID=$(curl -s -H "X-aws-ec2-metadata-token: $_CLOUD_TOKEN" http://169.254.169.254/latest/dynamic/instance-identity/document | grep imageId | cut -d\" -f4)
+    _CI_IP=$(curl -s -H "X-aws-ec2-metadata-token: $_CLOUD_TOKEN" http://169.254.169.254/latest/meta-data/local-ipv4)
     _CLOUD_PROVIDER=AWS
     _KUBE_NODE_NAME="$_CLOUD_INSTANCE_ID"
 
     useradd pipeline
+
     cp -r /home/ec2-user/.ssh /home/pipeline/.ssh
     chown -R pipeline. /home/pipeline/.ssh
     chmod 700 .ssh
@@ -327,43 +447,63 @@ _KUBE_LOG_ARGS="--logtostderr=false --log-dir=$_KUBELET_LOG_PATH"
 _KUBE_NODE_NAME="${_KUBE_NODE_NAME:-$(hostname)}"
 _KUBE_NODE_NAME_ARGS="--hostname-override $_KUBE_NODE_NAME"
 
+_KUBE_PULL_TIMEOUT="@kube_pull_deadline@"
+if [ -z "$_KUBE_PULL_TIMEOUT" ] || [[ "$_KUBE_PULL_TIMEOUT" == "@"*"@" ]]; then
+  _KUBE_PULL_TIMEOUT="10m"
+fi
+_KUBE_OTHER_ARGS="--image-pull-progress-deadline $_KUBE_PULL_TIMEOUT"
+_KUBE_OTHER_ARGS="--enable-cadvisor-json-endpoints $_KUBE_OTHER_ARGS"
+
 # FIXME: use the .NodeRegistration.KubeletExtraArgs object in the configuration files
 _KUBELET_INITD_DROPIN_PATH="/etc/sysconfig/kubelet"
 rm -f $_KUBELET_INITD_DROPIN_PATH
 
-## FIXME: shall be moved to the preferences
-_KUBE_RESERVED_RATIO=5
-_KUBE_RESERVED_MIN_MB=500
-_KUBE_RESERVED_MAX_MB=2000
-_KUBE_NODE_MEM_TOTAL_MB=$(awk '/MemTotal/ { printf "%.0f", $2/1024 }' /proc/meminfo)
-_KUBE_NODE_MEM_RESERVED_MB=$(( $_KUBE_NODE_MEM_TOTAL_MB * $_KUBE_RESERVED_RATIO / 100 / 2 ))
-_KUBE_NODE_MEM_RESERVED_MB=$(( $_KUBE_NODE_MEM_RESERVED_MB > $_KUBE_RESERVED_MAX_MB ? $_KUBE_RESERVED_MAX_MB : $_KUBE_NODE_MEM_RESERVED_MB ))
-_KUBE_NODE_MEM_RESERVED_MB=$(( $_KUBE_NODE_MEM_RESERVED_MB < $_KUBE_RESERVED_MIN_MB ? $_KUBE_RESERVED_MIN_MB : $_KUBE_NODE_MEM_RESERVED_MB ))
-_KUBE_RESERVED_ARGS="--kube-reserved cpu=300m,memory=${_KUBE_NODE_MEM_RESERVED_MB}Mi,ephemeral-storage=1Gi"
-_KUBE_SYS_RESERVED_ARGS="--system-reserved cpu=300m,memory=${_KUBE_NODE_MEM_RESERVED_MB}Mi,ephemeral-storage=1Gi"
+KUBE_RESERVED_CPU="300m"
+KUBE_RESERVED_DISK="1Gi"
+KUBE_RESERVED_MEM="@KUBE_RESERVED_MEM@"
+if [ ! "$KUBE_RESERVED_MEM" ] || [[ "$KUBE_RESERVED_MEM" == "@"*"@" ]]; then
+    KUBE_RESERVED_MEM="250Mi"
+fi
+
+SYSTEM_RESERVED_CPU="300m"
+SYSTEM_RESERVED_DISK="1Gi"
+SYSTEM_RESERVED_MEM="@SYSTEM_RESERVED_MEM@"
+if [ ! "$SYSTEM_RESERVED_MEM" ] || [[ "$SYSTEM_RESERVED_MEM" == "@"*"@" ]]; then
+    SYSTEM_RESERVED_MEM="250Mi"
+fi
+
+_KUBE_RESERVED_ARGS="--kube-reserved cpu=${KUBE_RESERVED_CPU},memory=${KUBE_RESERVED_MEM},ephemeral-storage=${KUBE_RESERVED_DISK}"
+_KUBE_SYS_RESERVED_ARGS="--system-reserved cpu=${SYSTEM_RESERVED_CPU},memory=${SYSTEM_RESERVED_MEM},ephemeral-storage=${SYSTEM_RESERVED_DISK}"
 _KUBE_EVICTION_ARGS="--eviction-hard= --eviction-soft= --eviction-soft-grace-period= --pod-max-pids=-1"
 _KUBE_FAIL_ON_SWAP_ARGS="--fail-swap-on=false"
 
-echo "KUBELET_EXTRA_ARGS=$_KUBE_NODE_INSTANCE_LABELS $_KUBE_LOG_ARGS $_KUBE_NODE_NAME_ARGS $_KUBE_RESERVED_ARGS $_KUBE_SYS_RESERVED_ARGS $_KUBE_EVICTION_ARGS $_KUBE_FAIL_ON_SWAP_ARGS" >> $_KUBELET_INITD_DROPIN_PATH
+_KUBELET_VERSION=$(kubelet --version | cut -f2 -d'.')
+if (( "$_KUBELET_VERSION" >=  19)); then
+  _KUBE_OTHER_ARGS="$_KUBE_OTHER_ARGS --feature-gates CSIMigration=false"
+fi
+
+echo "KUBELET_EXTRA_ARGS=$_KUBE_NODE_INSTANCE_LABELS $_KUBE_LOG_ARGS $_KUBE_NODE_NAME_ARGS $_KUBE_RESERVED_ARGS $_KUBE_SYS_RESERVED_ARGS $_KUBE_EVICTION_ARGS $_KUBE_FAIL_ON_SWAP_ARGS $_KUBE_OTHER_ARGS" >> $_KUBELET_INITD_DROPIN_PATH
 chmod +x $_KUBELET_INITD_DROPIN_PATH
 
 systemctl enable docker
 systemctl enable kubelet
 systemctl start docker
 
-for _KUBE_SYSTEM_POD_FILE in $_DOCKER_SYS_IMGS/*.tar; do
-  docker load -i $_KUBE_SYSTEM_POD_FILE
-done
-rm -rf $_DOCKER_SYS_IMGS
+if [ "$skip_system_images_load" != "true" ]; then
+  for _KUBE_SYSTEM_POD_FILE in $_DOCKER_SYS_IMGS/*.tar; do
+    docker load -i $_KUBE_SYSTEM_POD_FILE
+  done
+  rm -rf $_DOCKER_SYS_IMGS
+fi
 
 kubeadm join --token @KUBE_TOKEN@ @KUBE_IP@ --discovery-token-unsafe-skip-ca-verification --node-name $_KUBE_NODE_NAME --ignore-preflight-errors all
 systemctl start kubelet
 
 update_nameserver "$nameserver_post_val" "infinity"
 
+@custom_script_post@
+
 if [[ $FS_TYPE == "btrfs" ]]; then
-  _API_URL="@API_URL@"
-  _API_TOKEN="@API_TOKEN@"
   _MOUNT_POINT="/ebs"
   _FS_AUTOSCALE_PRESENT=0
   _CURRENT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
@@ -404,7 +544,9 @@ EOL
   fi
 fi
 
-if check_installed "nvidia-smi"; then
+enable_emergency_termination_service
+
+if check_gpu_available; then
   cat >> /etc/rc.local << EOF
 nvidia-persistenced --persistence-mode
 nvidia-smi
@@ -423,7 +565,14 @@ if [[ ! -z "${_PRE_PULL_DOCKERS}" ]] && [[ "${_PRE_PULL_DOCKERS}" != "@"*"@" ]] 
   IFS=',' read -ra DOCKERS <<< "$_PRE_PULL_DOCKERS"
   for _DOCKER in "${DOCKERS[@]}"; do
     _REGISTRY="${_DOCKER%%/*}"
-    docker login -u "$_API_USER" -p "$_API_TOKEN" "${_REGISTRY}"
+    for i in $(seq 1 10); do
+        echo "Trying $_REGISTRY"
+        docker login -u "$_API_USER" -p "$_API_TOKEN" "${_REGISTRY}"
+        _r=$?
+        [ $_r -eq 0 ] && break
+        sleep 3
+    done
+    [ $_r -ne 0 ] && echo "Cannot login to $_REGISTRY" && continue
     docker pull "$_DOCKER"
   done
 fi
